Remove cycle dependencies
