+++ tests/tests/lib.rs
@@ -10,7 +10,9 @@ mod contract;
 mod dap;
 mod debugger;
 mod deployment;
+mod fee_collection_contract;
 mod health;
+
 mod helpers;
 mod messages;
 mod metrics;
